+++ tests/testthat/test_auk-unique.r
@@ -1,4 +1,5 @@
 context("auk_unique")
+library(dplyr)
 
 test_that("auk_unique removes duplicates", {
   ebd <- system.file("extdata/zerofill-ex_ebd.txt", package = "auk") %>%
@@ -38,6 +39,23 @@ test_that("auk_unique removes duplicates in sampling file", {
   expect_equal(class(ebd), class(ebd_unique))
 })
 
+test_that("auk_unique combines ids correctly", {
+  ebd <- system.file("extdata/zerofill-ex_sampling.txt", package = "auk") %>%
+    read_sampling(unique = FALSE)
+  ebd_unique <- auk_unique(ebd, checklists_only = TRUE)
+  
+  ebd_g <- ebd %>% 
+    filter(!is.na(group_identifier)) %>% 
+    group_by(group_identifier) %>% 
+    arrange(sampling_event_identifier) %>% 
+    summarize(cid = paste(sampling_event_identifier, collapse = ","),
+              oid = paste(observer_id, collapse = ",")) %>% 
+    inner_join(ebd_unique, by = "group_identifier")
+  
+  expect_equal(ebd_g$cid, ebd_g$sampling_event_identifier)
+  expect_equal(ebd_g$oid, ebd_g$observer_id)
+})
+
 test_that("auk_unique throws error for invalid input", {
   ebd <- system.file("extdata/zerofill-ex_ebd.txt", package = "auk") %>%
     read_ebd(unique = FALSE)
@@ -48,4 +66,5 @@ test_that("auk_unique throws error for invalid input", {
   expect_error(auk_unique(ebd, group_id = "abc"))
   expect_error(auk_unique(ebd, checklist_id = "abc"))
   expect_error(auk_unique(ebd, species_id = "abc"))
+  expect_error(auk_unique(ebd, observer_id = "abc"))
 })